Added filter for strings with commas inside of it
Sometimes we want to parse information that contains commas inside the cell, so we can't split that information into two or more cells. Instead, we filter those splitted erroneously and join them again.

For example: the string "My Company, Inc." is the kind of data that can be found in a cell, but should not be split. This new function will detect the two parts `"My Company` and ` Inc"` and join them again using the comma.